Refactored identifier names
