+++ scripts/extract/sports-game-odds/sports.js
@@ -3,20 +3,16 @@
  * @fileoverview Test the query to retrieve sports data.
  */
 import 'dotenv/config';
-import * as Sports from "../../scripts/api/sports-game-odds/sports.js" ; 
+import * as Sports from "../../api/sports-game-odds/sports.js" ; 
 
 
-describe("Sports Game Odds - Retrieve sport data", function () {
+// Query Sports
+await Sports.getSports().then((sports) => {
 
-  it("Retrieve sport data", async function () {
+  sports.data.forEach((sport) => {
 
-    // Query Sports
-    await Sports.getSports().then((response) => {
-
-      // Check if the sports are not undefined
-      expect(response.data).not.toBeUndefined();
-
-    });
+    // Check if the sports are not undefined
+    console.log(sport);
 
   });
